+++ src/actions/build.rs
@@ -40,18 +40,16 @@ impl BuildAction<'_> {
                 Error::FailedToParseArgument("contracts_names".to_string()).print_and_die()
             });
 
+        utils::validate_external_contracts(self.project, &contracts);
+
         for contract in contracts {
-            let module_name = match self.project.is_workspace() {
-                true => contract.module_name(),
-                false => contract.crate_name(self.project),
-            };
-            let build_contract = format!("{}_build_contract", module_name);
+            let host_crate = contract.host_crate(self.project);
+            let build_contract = format!("{}_build_contract", host_crate.name);
             command::cargo_build_wasm_files(
                 self.project.project_root(),
-                &contract.struct_name(),
-                &module_name,
-                self.project.is_workspace(),
-                contract.module_crate_name(self.project),
+                &contract.odra_module_value(self.project),
+                &host_crate,
+                self.project.is_cargo_workspace(),
             );
 
             let source = paths::wasm_path_in_target(&build_contract, self.project.project_root());